==> 1. Checkpoint.ipynb <==
pip install pandas

**For EDA we use**
- Pandas -> used for data frame applications
- Numpy  -> Numerical python/ Numerical analysis. All math applications
- Matplitlib -> used for plotting
- Seaborn -> used for plot
- Plotly, Bokhe  -> used for plot

**Machine Learning**
- Sickit-learn -> sklearn

**Deep Learning**
- Tensorflow -> Google developed
- Keras -> Google
- Pytorch -> Facebook

**NLP(Natural Language Processing)**
- NLTK -> Natural Language Tool Kit
- Scipy

**BERT (Bi-directional Encoded Representation Transformers)**
- Transformers -> Developed by Hugging Face take over by Google

**Google Models**
- Specific Google Packages

**Azure ML**
- Azure Packages

**GenAI**
- Gemini AI -> Google
- ChatGPT -> OpenAI

**LangChain**
- LangChain packages

$Step$-$1$

**Create a DataFrame**

import pandas as pd
names=['Zoro', 'Luffy','Akaza']
pd.DataFrame() # Empty dataframe

$Step$-$2$ 

**Provide the Data**

names=['Zoro', 'Luffy','Akaza']
pd.DataFrame(names)

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
pd.DataFrame(names,age) # It is taking age as index

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
pd.DataFrame(zip(names,age))
# first positional argument: data
# second positional argument: index

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
pd.DataFrame(zip(names,age,city))

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
pd.DataFrame(zip(names,age),city)

$Step$-$3$:

**Provide the Column names**

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
pd.DataFrame(zip(names,age),
             columns=['Names','Age']) # By giving columns the name is changing

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city),
             columns=['Names','Age','City'])

type(df)

$Step$-$4$:
**Add new Column**

- If you want to add new column 
- First you need to check the no of rows in already existed data
- For Eg: in above DataFrame has 3 rows
- You need to create a new list with 3 rows
- That list equate to the DataFrame

job=['DS','DE','CS']
df['JOB']=job # Giving a new column
df

$Step$-$5$:

**Update Existing Column**

df['Age']=[35,40,50] # Changing existed column
df

$Step$-$6$:

**Change the Index**

df

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

**Note**
- No of list == No of columns
- No of values in list == No of rows
- When you create the data all the list will be in same length

$Step$-$7$:

**Shape**

- No of rows & No of columns

df.shape # Matrix format --> 3 rows, 3 columns

print(f"The no of rows ={df.shape[0]}")

$Step$-$8$:

**Drop the Column**

- In order to drop a column require 3 arguments
    - Column name
    - Axis: 
        - Axis=0 --> **represent rows**     
        - Axis=1 --> **represent columns**
    - Inplace: Dropping column means modifying the DataFrame
        - This modification can be saved in **same variable (inplace=True)** or different variable

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

df.drop('City',axis=1,inplace=True)
df

df1=df.drop('Age',axis=1)
df # if inplace=True is not there then the column will not change in that variable

df1

$Step$-$9$:

**Multiple Columns Drop**

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

df.drop(['Age','City'], axis=1,inplace=True)
df

$Step$-$10$:

**Drop The Rows**

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

df.drop('A',axis=0,inplace=True)
df

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

df.drop(['A','B'],axis=0,inplace=True)
df

$Step$-$11$:

**How to save the DataFrame**

- Generally DataFrames are 2 types:
    - csv (Comma Separated Value)
    - excel

names=['Zoro', 'Luffy','Akaza']
age=[20,25,30]
city=['hyd','blr','chennai']
df=pd.DataFrame(zip(names,age,city), index=['A','B','C'],
             columns=['Names','Age','City'])
df

# File name=data
# Extension: .csv
df.to_csv('data.csv')
df.to_csv('data1.csv',index=False)

df.to_excel('data2.xlsx',index=False)

$For$ $Article$

import pandas as pd
names=['Lohith','Hari','Bunny']
pd.DataFrame(names)

import pandas as pd
names=['Lohith','Hari','Bunny']
age=[20,25,19]
pd.DataFrame(zip(names,age),columns=['Name','Age']) # It is taking age as index

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny'],
'Age':[20,25,19]}
df = pd.DataFrame(dict)
df

df['Job']=['DE','WD','DS']
df

d1={'Job':['DE','WD','DS']}
pd.DataFrame(zip(dict,d1))

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny'],
'Age':[20,25,19],
'City':['Hyd','Chennai','USA']}
df = pd.DataFrame(dict)
# Add new column to the above df
job=['Data Engineer','Software Developer','Data Scientist']
df['Job']=job # Assigning Job as column name and list as values
df

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny','Venkat Rao'],
'Age':[20,25,19,20],
'City':['Hyd','Chennai','USA','Hyd'],
'Job':['Data Engineer','Software Developer','Data Scientist','Programmer']}
df = pd.DataFrame(dict)
df['Age']=[25,30,20,19]
df

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny','Venkat Rao'],
'Age':[20,25,19,20], # initial age values
'City':['Hyd','Chennai','USA','Hyd'],
'Job':['Data Engineer','Software Developer','Data Scientist','Programmer']}
df = pd.DataFrame(dict,index=['A','B','C','D'])
df

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny','Venkat Rao'],
'Age':[20,25,19,20], # initial age values
'City':['Hyd','Chennai','USA','Hyd'],
'Job':['Data Engineer','Software Developer','Data Scientist','Programmer']}
df = pd.DataFrame(dict,index=['A','B','C','D'])

# Dropping City column
df.drop('City',axis=1,inplace=True)
df

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny','Venkat Rao'],
'Age':[20,25,19,20], # initial age values
'City':['Hyd','Chennai','USA','Hyd'],
'Job':['Data Engineer','Software Developer','Data Scientist','Programmer']}
df = pd.DataFrame(dict,index=['A','B','C','D'])

# Dropping City, Age columns
df.drop(['Age','City'],axis=1,inplace=True)
df

import pandas as pd
dict={'Names':['Lohith','Hari','Bunny','Venkat Rao'],
'Age':[20,25,19,20], # initial age values
'City':['Hyd','Chennai','USA','Hyd'],
'Job':['Data Engineer','Software Developer','Data Scientist','Programmer']}
df = pd.DataFrame(dict,index=['A','B','C','D'])

# Dropping Age, City columns
df.drop(['A','C'],axis=0,inplace=True)
df
