==> 2. Checkpoint-2.ipynb <==
**How to read a DataFrame**

- To read any data we need: **Location, File name, Extension**

import pandas as pd

pd.read_csv('data1.csv')

pd.read_excel('data2.xlsx')

$Task$

- Download visa & bank dataset

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df

bank_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\bank - bank.csv")
bank_df

**Create DatFrame using Dictionary**

names=['Zoro','Luffy','Robin']
age=[25,20,30]
pd.DataFrame(zip(names,age),columns=['Names','Age'])

dict1={'Names':['Zoro','Luffy','Robin'],
      'Ages':[25,20,30]}
dict1

pd.DataFrame(dict1)
#Keys become columns
# Values become rows
# No need to provide column names separately

dict2={'Names':['Zoro','Luffy','Nami'],
      'Ages':[20,25,20],
      'City':['Hyd','Blr','Chennai']}
pd.DataFrame(dict2,index=[1,2,3])
