==> 3. Visa Dataset Checkpoint.ipynb <==
**Import required packages**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

**Read the Data**

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df

$HEAD$

- Prints top 5 rows

visa_df.head()

visa_df.head(3)

$TAIL$

- Prints last 5 rows

visa_df.tail()

$SHAPE$

- Prints no of rows & columns

visa_df.shape

print(f"The no of rows = {visa_df.shape[0]}")
print(f"The no of columns = {visa_df.shape[1]}")

$SIZE$

- No of rows * No of Columns

visa_df.size

25480*12

$COLUMNS$

visa_df.columns

$D-TYPES$

visa_df.dtypes

# object means Categorical
# int, float means Numerical

- visa_df type is --> **DataFrame**
- visa_df values is --> **Series**
- Series objects can be converted into dictionary
- After thet we can use dictionary methods

type(visa_df)

type(visa_df.dtypes)

**Task**

- Create a categorical list
- Create numerical column list

**Idea**

- Convert series type into dict
- Iterate through loop
- Apply the condition get the result in list

dtypes=dict(visa_df.dtypes)
dtypes

cat_cols=[key for key,value in dtypes.items() if value=='object']
num_cols=[key for key,value in dtypes.items() if value!='object']

$Select$-$data$-$types$

- Select data types will give a DataFrame of desired data types
- The result also a DataFrame
- All above options like head,tail,columns

visa_df.select_dtypes(include='object')

visa_df.select_dtypes(include='object').columns

visa_df.select_dtypes(exclude='object').columns

$Is$-$Null$

- It is used to identify any missing values are available
- It means we are asking a qn to the computer
- It will give True or False
- True --> Yes value is missed
- False --> No value is not missed

**Note**
- If we see bound method means add the brackets

visa_df.isnull()

visa_df.isna()

visa_df.isnull().sum() # No null is available so sum=0

$DROP$ $Duplicates$

- If any duplicate rows are there then we can drop it

visa_df.drop_duplicates()

visa_df.drop_duplicates().all()

$Info$

visa_df.info()

$Take$

- It is used to extract the particular index of data
- It has one parameter --> axis
    - axis=0 --> rows
    - axis=1 --> columns
- By default axis=0 is there

visa_df.take([100,200,300])

visa_df.take([100,200,300],axis=1)

visa_df.take([2,5,7],axis=1)

visa_df.take([100:105]) # Not working

visa_df.take([100,200,300]).take([2,5,7],axis=1)

**Drwaback**

- Indices of rows not work
- At a time extract rows & columns is not possible in single take

$ILOC$

#visa_df.iloc[<rows>,<columns>]
#visa_df.iloc[<start:end>,<start:end>]

visa_df.iloc[[100,200,300],[2,5,7]]

visa_df.iloc[[100,200,300]] #Same as take

visa_df.iloc[100:105,[2,5,7]]  # While giving indices no brackets will be given

visa_df.iloc[100:105,2:5]

visa_df.iloc[100:105,[2]]

visa_df.iloc[100]

visa_df.iloc[[100],0:]

**Note**

- No bracket then series , with [] then DataFrame

visa_df.iloc[:,[5]]

type(visa_df.columns)

visa_df.columns

cols=list(visa_df.columns)
cols.index('no_of_employees')

$loc$

- It will take directly the column names

visa_df.loc[[100,200,300],['no_of_employees']]
