==> 8. Categorical to Numerical.ipynb <==
**Categorical to Numerical**

- It is very important to convert categorical data to numerical 
- Bcz in ML models the input is in form of numbers
- ML models developed by maths, So passing text data will not works
- We have different types of methods are there to convert categorical to numerical
    - Map method
    - np.where
    - One hot encoder
    - Label encoder

**Import the packages**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

**Read the data**

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df.head(2)

**Extract Categorical columns**

cat=visa_df.select_dtypes(include='object').columns
cat

**Map method**

- Get unique labels for each column
- Create a dict by providing a values to each label
- Eg: case_status having two labels, we are providing 2 values
    - Assign 0 for certified
    - Assign 1 for denied

visa_df['case_status'].unique()

d={'Denied':1,'Certified':0} # According to alphabetical order assign values
d

visa_df['case_status_new'] = visa_df['case_status'].map(d) # If get NaN then again read the data

visa_df

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df.head(2)

d={'Denied':1,'Certified':0}
visa_df['case_status']=visa_df['case_status'].map(d)
visa_df

d={'N':0,'Y':1}
visa_df['has_job_experience'] = visa_df['has_job_experience'].map(d)
#visa_df['requires_job_training'] = visa_df['requires_job_training'].map(d)
visa_df

**Task**

- Read the data
- Apply for loop, iterate through columns
- Get unique labels of each labels
- Get count of unique labels
- Create a dict
- Apply into map

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
for i in list(visa_df.select_dtypes(include='object').columns):
    labels = list(visa_df[i].unique())
    values = [i for i in range(len(labels))]
    d = dict(zip(labels,values))
    visa_df[i] = visa_df[i].map(d)
visa_df

**Label Encoder**

- It is a package under **scikit-learn**
- Scikit learn is a package which is heart of ML
- Inside scikit learn we have a method called **preprocessing**
- Under preprocessing we have labelEncoder
- Any **sklearn** package have 3 steps
    - Read the package
    - Save the package
    - Apply the fit transform

# Read the data
visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df.head(2)

# Step-1 Read the package
from sklearn.preprocessing import LabelEncoder
# Step-2 Save the package
le=LabelEncoder()
# Step-3 Apply Fit transform
visa_df['case_status'] = le.fit_transform(visa_df['case_status'])
visa_df

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
for i in visa_df.select_dtypes(include='object').columns:
    visa_df[i] = le.fit_transform(visa_df[i])
visa_df

**Inverse Transform**

- In order to see inverse of labels we need to perform individually again for each columns

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
# Step-1 Read the package
from sklearn.preprocessing import LabelEncoder
# Step-2 Save the package
le=LabelEncoder()
# Step-3 Apply Fit transform
visa_df['case_status'] = le.fit_transform(visa_df['case_status'])
print(visa_df['case_status'].values) # values is used for getting values in list format
print(le.inverse_transform(visa_df['case_status']))

**Label Encoder will be used most of the time**

**np.where method**

- It is applicable only for binary
- It takes 3 arguments
    - Condition
    - True value
    - False value

$Scenario$

- Eg: case_status = 'Certified' then replace with 0 otherwise 1

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
con = visa_df['case_status'] == 'Certified'
visa_df['case_status'] = np.where(con,0,1)
visa_df

**One hot Encoder**

- One hot means if one is ON another one is OFF
- On --> 1 and OFF --> 0
- Eg: 'case_status' has 2 unique labels('Denied', 'Certified')
- If apply this on case_status it will create 2 extra columns like case_status_denied & case_status_certified

| case_status|case_status_denied|case_status_certified|
|---------|--|--|
| Denied|1|0|
| Certified|0|1|

**Advantages**

- In ML we have one imp property is **Independent of variables**
- ML models expect that variables are **independent each other**
- One hot encoder creating new columns, which are **independent each other**
- Independent variables means both var have **90 degree phase shift**
- 90 degree phase shift --> perpendicular each other --> orthogonal each other 

**Diasdvantages**

- One hot encoder creates new columns = no of unique labels
- If no of unique labels increased then new columns also increased
- The more columns lead more data
- More data --> complexity increased --> more computation power require --> time increased
- The entire process named as **Curse of dimensionality**

**pd.get_dummies**

- Read data
- Choose one column
- Apply pd.get_dummies

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
pd.get_dummies(visa_df['case_status'],dtype='int')

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
pd.get_dummies(visa_df, dtype='int')

**Note:**
Always drop the id column

visa_df=pd.read_csv(r"C:\Users\Lenovo\Python NareshIT\Data files\Visadataset - Visadataset.csv")
visa_df.drop('case_id',axis=1,inplace=True)
pd.get_dummies(visa_df, dtype='int')

- Generally we prefer Label Encoder
- One hot encoder is useful when categorical data has less unique labels
